# backend/auth-service/src/main/java/com/procurelens/auth/dto/LoginRequest.java
package com.procurelens.auth.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class LoginRequest {
    private String email;
    private String password;
    private String role;

    public LoginRequest(String email, String password) {
        this.email = email;
        this.password = password;
    }
}